keep TechComm
